etldjango/requirements.txt
anyio==2.2.0
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextvars==2.4
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.5.0
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-pandas==0.6.4
django-storages==1.9.1
djangorestframework==3.12.2
entrypoints==0.3
google-api-core==1.26.3
google-auth==1.29.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-error-reporting==1.1.2
google-cloud-logging==2.3.1
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googlemaps==4.4.5
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
gunicorn==20.0.4
idna==2.10
immutables==0.15
importlib-metadata==3.7.3
install==1.3.4
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.2
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
libcst==0.3.18
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
#pkg-resources==0.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
rsa==4.7.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
tabula-py==2.2.0
terminado==0.9.4
testpath==0.4.4
tomlkit==0.7.0
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
